refactored registration
