+++ components/registration/ScanModal.js
