+++ components/registration/index.js
